feat: moved all the types to the types.ts file
