+++ packages/crayons-core/src/utils/types.ts
@@ -1,0 +1,17 @@
+export type TagVariant = 'standard' | 'avatar';
+
+export type DropdownVariant = 'standard' | 'icon' | 'avatar';
+
+export type PopoverPlacementType =
+  | 'top-start'
+  | 'top'
+  | 'top-end'
+  | 'left-start'
+  | 'left'
+  | 'left-end'
+  | 'right-start'
+  | 'right'
+  | 'right-end'
+  | 'bottom-start'
+  | 'bottom'
+  | 'bottom-end';
